Copy GCE instance fields into both Instance and InstanceTemplate
The hcl2 library doesn't handle embedded types well,
so rather than trying to detect it with reflection I figured it was easier to just get rid of the one case of it.
This copies the fields from the embedded type into both Instance and InstanceTemplate, updating the receiver functions to instead return the value for the field that was being modified.