Rename test names to make it easier to search for failing tests
This is very minor...

GH Action's search functionality is not case sensitive.
In Travis CI, one could search for `FAIL:` and be taken to relevant test failures.
In GHA, these tests were also in the search results making it harder to find the actual tests that are failing.

This renames the tests so that searching for `FAIL:` (case insensitive) will only return failing tests