Allow CI builds to build a tagged version
This is working towards "hands off" release builds, where we just tag
the release.  When CI=1 and the git sha we are building _exactly_
matches a tag, we build that version.  We enforce that version.go must
have previously been updated to match the tag.

We also refactor out the duplicated code into a shared script (tools/get_version.sh)