validate_cluster: Create node to instance group mapping to get pod instance group
In the ValidationError struct, there is a field to identify the instance group to
which the ValidationError is associated with.

For pod related ValidationErrors, it is not straightforward to identify the
instance group to which the pod is associated with.

To acheive this, we create a node to instance group mapping in ValidateNodes.
This node to instance group mapping is used in collectPodFailures to identify the
pod instance group by using the pod's hostIp field.

We don't associate system-cluster-critical pods to instance groups as those pod
failures are cluster wide