Precreate the kops-controller DNS name
We're seeing test failures in prow because nodes are not able to resolve the kops-controller dns record.

Judging by timestamps, dns-controller adds the record after it is first queried by nodeup.
The negative TTL is long enough that the cluster doesn't validate within the alotted time.
Rather than increasing the validation timeout I think its better to precreate the DNS record the same way we do for the other records.