Fix MetricsGrabber skip regex condition
The release/latest version marker will stay on 1.23.0-alpha.0 until another 1.23 pre-release is tagged which isnt for a while.
This will skip the MetricsGrabber test for that tag which has been fixed in master. Once a new 1.23 pre-release has been tagged we can remove this.

Alternatively we migrate all jobs off of the release/latest version marker since it isn't particularly valuable.