Note for the metrics configuration for etcd
As etcd is running directly as a service on the master nodes it is easy to overlook that different ports are needed to expose metrics for multiple etcd clusters on the master nodes. It can be easy to forget this if you otherwise just deal with containers and pods... 😅  A reminder and clarification doesn't hurt.