Use a fixed set of pruning kinds
We don't want to prune some kinds that are considered dangerous
(CustomResourceDefinitions, which delete all the instances of the CRD)
and Namespaces (which delete all the objects in that Namespace).

In addition, having a hard-coded list means we don't have to remember
to manually specify the kind if we delete an object from the manifest
(as long as it's in the known set).

We can in future support specifying additional kinds, but we can cross
that when we need it.