pruning: Allow specification of FieldSelector
FieldSelector allows for filtering by name (as well as a few other
attributes).  Though we aren't using this today, filtering by name is
something we've discussed as being valuable in future, and by putting
it in the API now we avoid version skew problems should we need to
introduce it in future.