addons: support for kopeio-networking addon
Adding support for the kopeio-networking addon.

We load the operator manifest (which changes rarely) from the channels
directory for now.  We follow the same structure as operators
themselves use so that we can support other backends in future.

The channels file includes the current versions of the operators.

During cluster creation, we create these additional objects.