e2e: fix test flakes where we specify a non-schedulable zone
If we specify a zone to the e2e tests, it will try to pre-provision
volumes into it for the inline tests.  We then will be unable to
attach a volume to test pods if we don't have (normal) nodes in that
zone.

Pick a zone that is schedulable for test pods.