Fix OpenStack SecurityGroupRule/LB with IPv6 CIDR
When creating security group rule, the default EtherType is IPv4.
Currently, AdminAccess default includes IPv6 CIDR, in this case
EtherType should be changed to IPv6.

Similar issue also happens with load balancer creation as the VIP is
IPv4.

This commit fixed this issue by checking if the CIDR is IPv6 before
creating the rule, if it is, EtherType will be changed to IPv6. When
creating load balancer, if allowed CIDR doesn't match VIP type, remove
it from the list.