WIP: Support channels over OCI
We want to support richer mirroring for channels, into the official k8s distribution locations.

However, our channels currently update in place, and this doesn't seem
like a great match for the http-based distribution mechanisms (which
both expect artifacts to be immutable and don't support listing).

Instead, we support fetching the files via a container (OCI registry);
we still treat them as immutable but OCI makes it easy to list tags.