Be more specific when filtering OS instance ports
This adds an additional filter for ports attached to an instance, so
that only ports tagged with the cluster name will be considered.

Basically, this enables adding ports to the server independent from the
provisioning process.

To not break clusters provisioned with older kOps versions, when there
are no tagged ports found it will still consider all the ports of an
instance.