Retry creating failed OpenStack servers
This aims to improve the experience when creating openstack servers and
they run into issues during scheduling, which is not covered by the
create API request.

So after creating the instance kops waits for it to become ACTIVE and if
not, tries to reprovision the instance by deleting the failed instance
and creating a new one.

If the last attempt was still not successful, erroneous instances will
persist to allow further investigation, and the task will fail, which
will ultimately fail the update call.