gce: Don't reconcile routes when running with "gce" networking.
If running with GCE "native" networking, we do not need the route
controller (and it causes problems); we turn it off by setting
--configure-cloud-routes=false.

In general we do not need the gkenetworkparams controller (and it
complains about missing CRDs).  We will turn it off in future, but it
isn't in the images we are using currently.