OpenStack: Build InstanceGroup despite missing IP
This will allow kOps to build an OpenStack InstanceGroup with a missing
fixed IP.

In essence this will get rid of `interface name X not found` errors,
when there are servers present which do not have an interface attached
or are in a state (e.g. `ERROR`) which does not allow this.