Add a mutex lock to 'awsCloudInstances' map
We're using terraform kops provider to manage our AWS kops clusters.
Time to time we hit a race condition with the stack trace points to
`awsup.NewAWSCloud` function, when writing to a concurrent map, that
maintains a map between regions and `AWSCloud` objects.

This PR changes this to variable so it belongs to a new type, that wraps
the map into its own struct where access is controlled by a mutex lock.

Let me know if that makes sense to you all.

Thanks for building this awesome project!