Don't get cluster --full when attempting leak cleanup
If a job's --down was interrupted, the state store can be in an inconsistent state.
This can cause `kops get cluster --full` to fail if certain additional objects are missing, yet `kops get cluster` would succeed.
We omit `--full` to ensure `kops get cluster` succeeds and leak cleanup is reattempted.