devcontainer: update go version, use features
From what I've read, we want the base image to be smaller, with more
of the added functionality in features.  We can build a bigger image
with the features, but that should be an image for devcontainers.

Also it seems that devcontainer tooling will cache the image with the
features - maybe giving similar results to if we had built a more
complete image.