Provide default region to route53 client
This has similar behavior to awsup.ValidateRegion where we dont have access to a kops.Cluster to determine its region.

This is only needed in the dnsprovider client initialized in the kops CLI for creating the placeholder records - it wont be used on control plane nodes because LoadDefaultConfig will query IMDS for the node's region.

This isn't needed for all other usage of awsconfig.LoadDefaultConfig() because the region is known, for example in awsup.NewAWSCloud().

In any other partitions where us-east-1 is not appropriate, the user will already be specifying a region in their ~/.aws/config or via env var.