Make ASG Warmpool depend on ASG Lifecycle hook
Fixes: https://github.com/kubernetes/kops/issues/16582

When creating a new ASG with a Warmpool using Lifecycle hooks, the
instances that first join the Warmpool when the ASG is created could
come up before the Lifecycle hook is in effect.
This can lead to problems such as those instances not calling the
Lifecycle hook notification when pressed into service, causing the ASG
to terminate them approximately 10 mins after they've been performing
work in the cluster.

Setting a dependency on for the Warmpool to take into account any
Lifecycle hooks means that the Warmpool won't be created until the
hooks are ready.

Signed-off-by: Jim Barber <jim.barber@healthengine.com.au>