Tweak kube-apiserver timeouts
It looks like the kube-apiserver startup has slowed down a lot
(e.g. creating a lot of RBAC stuff), and this likely times out with
the current settings on smaller machines (as found by our
VMs-in-github-actions tests).

Increase the timeouts, aligning with k/k and using the newer healthcheck endpoints.