Remove some default CPU limits
Remove the default CPU limits defined for the aws-iam-authentication and
node-problem-detector DaemonSets.

This makes them behave the same as the other `cpuLimit` parameters for
the cluster that also do not have defaults.

As it was previously set up, if an administrator does not want CPU
limits defined for these DaemonSets, there was no way to define that via
the cluster spec.