fix kops get cluster for files
We were skipping tasks in `kops get assets`; but this
meant we would not discover assets for skipped tasks.

We are doing a dry-run here, so introduce a dry-run mode that
does not look for existing resources.  That is essentially as fast
as skipping the task, but means we visit all tasks.