fix: use kopsbase.Version instead of kopsbase.KOPS_RELEASE_VERSION
Normally they are the same, but for CI builds they are different,
and kopsbase.Version is the consistent one to use for CI builds.

We also need to be careful not to conflate the version with
the docker image tag; image tags cannot contain '+' characters,
but our CI versions do.  We replace '+' with '_' for image tags.

Co-authored-by: Ciprian Hacman <ciprian@hakman.dev>